Added i18n label lookup for toggles.
Unfortunately, there doesn't appear to be an API available that
internationalizes labels WITHOUT also wrapping them in a label tag. We
want the label text for toggles to be wrapped by a label tag, but the
actual text of the label should come AFTER the input.

For lack of a better solution, I have opted to call `FormBuilder.label`
but then strip the label tag fro the returned value via gsub.